+++ sertit/__meta__.py
@@ -17,4 +17,4 @@
 Sertit Utils library
 """
 
-__version__ = "1.45.0"
+__version__ = "1.45.1"